refactor: インプット読み込み処理をOIR用とEIR用に分割する
